+++ catalog/app/components/Logo/index.spec.tsx
@@ -12,26 +12,20 @@ describe('components/Logo', () => {
     mocked(useMediaQuery).mockClear()
   })
 
-  it('should render', () => {
-    const tree = renderer.create(<Logo />).toJSON()
+  it('should render squared logo', () => {
+    const tree = renderer.create(<Logo height="20px" width="20px" />).toJSON()
     expect(tree).toMatchSnapshot()
   })
 
-  it('should be short when forcedShort', () => {
-    const tree = renderer.create(<Logo forcedShort />).toJSON()
+  it('should render rectangular logo', () => {
+    const tree = renderer.create(<Logo height="30px" width="60px" />).toJSON()
     expect(tree).toMatchSnapshot()
   })
 
-  it('should be default in wide viewport', () => {
-    const tree = renderer.create(<Logo responsive />).toJSON()
-    expect(tree).toMatchSnapshot()
-  })
-
-  it('should be short in narrow viewport', () => {
-    // TODO: mock `window.matchMedia` instead of `M.useMediaQuery`
-    // NOTE: it means `xs === true`
-    mocked(useMediaQuery).mockImplementation(() => true)
-    const tree = renderer.create(<Logo responsive />).toJSON()
+  it.skip('should render custom logo', () => {
+    const tree = renderer
+      .create(<Logo src="https://example.com/example.png" height="10px" width="10px" />)
+      .toJSON()
     expect(tree).toMatchSnapshot()
   })
 })