+++ catalog/app/containers/NavBar/Help.tsx
@@ -1,3 +1,4 @@
+import * as FP from 'fp-ts'
 import * as React from 'react'
 import * as M from '@material-ui/core'
 import * as Lab from '@material-ui/lab'
@@ -30,67 +31,85 @@ const useStyles = M.makeStyles((t) => ({
   },
 }))
 
-const syntaxHelpRows = [
+type ExampleFormatter = (...syntaxParts: React.ReactNode[]) => React.ReactNode
+
+interface SyntaxHelpItem {
+  example: ExampleFormatter
+  syntax: string | string[]
+  title: string
+}
+
+interface SyntaxHelpNamespace {
+  namespace: string
+  rows: SyntaxHelpItem[]
+}
+
+const syntaxHelpRows: SyntaxHelpNamespace[] = [
   {
     namespace: 'Fields',
     rows: [
       {
-        example: 'comment:TODO',
+        example: (s) => <>{s}TODO</>,
         syntax: 'comment:',
         title: 'Package comment',
       },
       {
-        example: 'content:Hello',
+        example: (s) => <>{s}Hello</>,
         syntax: 'content:',
         title: 'Object content',
       },
       {
-        example: 'ext:*.fastq.gz',
+        example: (s) => <>{s}*.fastq.gz</>,
         syntax: 'ext:',
         title: 'Object extension',
       },
       {
-        example: 'handle:"user/*"',
+        example: (s) => <>{s}"user/*"</>,
         syntax: 'handle:',
         title: 'Package name',
       },
       {
-        example: 'hash:3192ac1*',
+        example: (s) => <>{s}3192ac1*</>,
         syntax: 'hash:',
         title: 'Package hash',
       },
       {
-        example: 'key:research*',
+        example: (s) => <>{s}research*</>,
         syntax: 'key:',
         title: 'Object key',
       },
       {
-        example: 'key_text:"research"',
+        example: (s) => <>{s}"research"</>,
         syntax: 'key_text:',
         title: 'Analyzed object key',
       },
       {
-        example: 'metadata:dapi',
+        example: (s) => <>{s}[2022-02-04 TO 2022-02-20]</>,
+        syntax: 'last_modified:',
+        title: 'Last modified date',
+      },
+      {
+        example: (s) => <>{s}dapi</>,
         syntax: 'metadata:',
         title: 'Package metadata',
       },
       {
-        example: 'size:>=4096',
+        example: (s) => <>{s}&gt;=4096</>,
         syntax: 'size:',
         title: 'Object size in bytes',
       },
       {
-        example: 'version_id:t.LVVCx*',
+        example: (s) => <>{s}t.LVVCx*</>,
         syntax: 'version_id:',
         title: 'Object version id',
       },
       {
-        example: 'package_stats\n  .total_files:>100',
+        example: (s) => <>{s}&gt;100</>,
         syntax: 'package_stats\n  .total_files:',
         title: 'Package total files',
       },
       {
-        example: 'package_stats\n  .total_bytes:<100',
+        example: (s) => <>{s}&lt;100</>,
         syntax: 'package_stats\n  .total_bytes:',
         title: 'Package total bytes',
       },
@@ -101,28 +120,32 @@ const syntaxHelpRows = [
     namespace: 'Logical operators and grouping',
     rows: [
       {
-        example: 'a AND b',
+        example: (s) => <>a {s} b</>,
         syntax: 'AND',
         title: 'Conjunction',
       },
       {
-        example: 'a OR b',
+        example: (s) => <>a {s} b</>,
         syntax: 'OR',
         title: 'Disjunction',
       },
       {
-        example: 'NOT a',
+        example: (s) => <>{s} a</>,
         syntax: 'NOT',
         title: 'Negation',
       },
       {
-        example: '_exists_: content',
+        example: (s) => <>{s} content</>,
         syntax: '_exists_:',
         title: 'Matches any non-null value for the given field',
       },
       {
-        example: '(a OR b)',
-        syntax: '()',
+        example: (s1, s2) => (
+          <>
+            {s1}a AND b{s2}
+          </>
+        ),
+        syntax: ['(', ')'],
         title: 'Group terms',
       },
     ],
@@ -132,46 +155,82 @@ const syntaxHelpRows = [
     namespace: 'Wildcards and regular expressions',
     rows: [
       {
-        example: 'config.y*ml',
+        example: (s) => <>config.y{s}ml</>,
         syntax: '*',
-        title: 'Zero or more characters, avoid leading * (slow)',
+        title: 'Zero or more characters, avoid leading * (slow)',
       },
       {
-        example: 'React.?sx',
+        example: (s) => <>React.{s}sx</>,
         syntax: '?',
         title: 'Exactly one character',
       },
       {
-        example: '/lmnb[12]/',
-        syntax: '//',
+        example: (s1, s2) => (
+          <>
+            {s1}lmnb[12]{s2}
+          </>
+        ),
+        syntax: ['/', '/'],
         title: 'Regular expression (slows search)',
       },
     ],
   },
 ]
 
-function Item({ item }) {
+const useExamplePartStyles = M.makeStyles((t) => ({
+  root: {
+    color: t.palette.text.hint,
+  },
+  syntax: {
+    color: t.palette.text.primary,
+  },
+}))
+
+interface ExamplePartProps {
+  syntax: string | string[]
+  example: ExampleFormatter
+}
+
+function Example({ syntax, example }: ExamplePartProps) {
+  const classes = useExamplePartStyles()
+  const syntaxParts = (Array.isArray(syntax) ? syntax : [syntax]).map((part) => (
+    <span className={classes.syntax}>{part}</span>
+  ))
+  return <span className={classes.root}>{example(...syntaxParts)}</span>
+}
+
+interface SyntaxProps {
+  syntax: string | string[]
+}
+function Syntax({ syntax }: SyntaxProps) {
+  return <>{Array.isArray(syntax) ? syntax.join('') : syntax}</>
+}
+
+interface ItemProps {
+  item: SyntaxHelpItem
+}
+
+function Item({ item }: ItemProps) {
   const t = M.useTheme()
   const sm = M.useMediaQuery(t.breakpoints.up('sm'))
 
   const { example, syntax, title } = item
   return (
     <M.Grid container>
-      <M.Grid item xs={6} sm={3}>
+      <M.Grid item xs={7}>
         <M.Typography variant="body2">
-          <Code>{syntax}</Code>
+          <Code>
+            {sm ? (
+              <Example syntax={syntax} example={example} />
+            ) : (
+              <Syntax syntax={syntax} />
+            )}
+          </Code>
         </M.Typography>
       </M.Grid>
       <M.Grid item xs>
         <M.Typography variant="body2">{title}</M.Typography>
       </M.Grid>
-      {sm && (
-        <M.Grid item xs={4}>
-          <M.Typography variant="body2">
-            <Code>{example}</Code>
-          </M.Typography>
-        </M.Grid>
-      )}
     </M.Grid>
   )
 }
@@ -184,24 +243,16 @@ const useItemsHeaderStyles = M.makeStyles((t) => ({
 }))
 
 function ItemsHeader() {
-  const t = M.useTheme()
-  const sm = M.useMediaQuery(t.breakpoints.up('sm'))
-
   const classes = useItemsHeaderStyles()
 
   return (
     <M.Grid container className={classes.root}>
-      <M.Grid item xs={6} sm={3}>
+      <M.Grid item xs={7}>
         <M.Typography variant="subtitle2">Command</M.Typography>
       </M.Grid>
       <M.Grid item xs>
         <M.Typography variant="subtitle2">Description</M.Typography>
       </M.Grid>
-      {sm && (
-        <M.Grid item xs sm={4}>
-          <M.Typography variant="subtitle2">Example</M.Typography>
-        </M.Grid>
-      )}
     </M.Grid>
   )
 }
@@ -227,9 +278,19 @@ function DocsExternalLink() {
   )
 }
 
-const normalizeSyntaxItem = (s) => s.replace(/\s/g, '')
+const normalizeSyntaxItem = (s: string | string[]) =>
+  FP.function.pipe(
+    s,
+    (s1: string | string[]) => (Array.isArray(s1) ? s1.join('') : s1),
+    (s2: string) => s2.replace(/\s/g, ''),
+  )
 
-export default function Help({ className, onQuery, ...props }) {
+type HelpProps = Partial<React.HTMLAttributes<HTMLDivElement>> & {
+  className: string
+  onQuery: (query: string) => void
+}
+
+export default function Help({ className, onQuery, ...props }: HelpProps) {
   const classes = useStyles()
 
   return (
@@ -254,16 +315,19 @@ export default function Help({ className, onQuery, ...props }) {
             <M.List className={classes.subList}>
               <ItemsHeader />
 
-              {rows.map((item) => (
-                <M.ListItem
-                  key={item.syntax}
-                  className={classes.item}
-                  button
-                  onClick={() => onQuery(normalizeSyntaxItem(item.syntax))}
-                >
-                  <Item item={item} />
-                </M.ListItem>
-              ))}
+              {rows.map((item) => {
+                const syntax = normalizeSyntaxItem(item.syntax)
+                return (
+                  <M.ListItem
+                    key={syntax}
+                    className={classes.item}
+                    button
+                    onClick={() => onQuery(syntax)}
+                  >
+                    <Item item={item} />
+                  </M.ListItem>
+                )
+              })}
             </M.List>
           </Lab.TreeItem>
         ))}