+++ catalog/app/embed/Dir.js
@@ -81,8 +81,8 @@ export default function Dir({
         label: 'Python',
         hl: 'python',
         contents: dedent`
-          import quilt3
-          b = quilt3.Bucket("s3://${bucket}")
+          import quilt3 as q3
+          b = q3.Bucket("s3://${bucket}")
           # list files
           b.ls("${path}")
           # download